base on bitmap
